agent-web/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
jinja2>=3.1.4
python-multipart>=0.0.9
pydantic>=2.8.2
sqlmodel>=0.0.21
aiosqlite>=0.20.0
psycopg2-binary>=2.9.0
langgraph>=0.2.34
langchain-community>=0.3.0
langchain-core>=0.3.0
chromadb>=0.5.5
pyyaml>=6.0.2
regex>=2024.7.24
langchain_gigachat>=0.3.12
langchain-tavily>=0.2.1
requests>=2.31.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0
